pyproject.toml
[build-system]
requires = ["setuptools>=77.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "humanbound"
version = "2.0.0"
description = "Humanbound — open-source AI agent red-team engine, SDK, and CLI."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    {name = "Humanbound", email = "hello@humanbound.ai"},
]
maintainers = [
    {name = "Demetris Gerogiannis", email = "hello@humanbound.ai"},
    {name = "Kostas Siabanis", email = "hello@humanbound.ai"},
]
keywords = [
    "ai",
    "ai-security",
    "red-team",
    "llm",
    "prompt-injection",
    "agent-security",
    "penetration-testing",
    "humanbound",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Environment :: Console",
    "Framework :: Pytest",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]

dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "requests>=2.32.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.0",
    "msal>=1.31.0",
    "pyperclip>=1.8.0",
]

[project.optional-dependencies]
engine = [
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "google-generativeai>=0.3.0",
    "websockets>=12.0",
]
firewall = [
    "humanbound-firewall>=0.2",
    "scikit-learn>=1.3.0",
    "torch>=2.0.0",
]
mcp = [
    "mcp>=1.2.0",
]
pytest = [
    "pytest>=7.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-timeout>=2.3",
    "ruff>=0.6",
    "mypy>=1.11",
    "build>=1.2",
    "twine>=5.0",
    "pre-commit>=3.8",
]

[project.scripts]
hb = "humanbound_cli.main:main"

[project.entry-points.pytest11]
hb = "humanbound_cli.pytest_plugin"

[project.urls]
Homepage = "https://github.com/humanbound/humanbound"
Documentation = "https://docs.humanbound.ai/"
Repository = "https://github.com/humanbound/humanbound"
Issues = "https://github.com/humanbound/humanbound/issues"
Changelog = "https://github.com/humanbound/humanbound/blob/main/CHANGELOG.md"
Community = "https://docs.humanbound.ai/community/"

[tool.setuptools.packages.find]
include = ["humanbound*", "humanbound_cli*"]

[tool.setuptools.package-data]
humanbound = ["py.typed"]
humanbound_cli = ["templates/*.html", "templates/*.svg", "py.typed"]

[tool.pytest.ini_options]
markers = [
    "hb: mark test as a Humanbound security test",
    "hb_category(name): mark test for a specific category",
]
timeout = 60
# tests/unit/test_*.py does `from conftest import ...` directly, so its
# directory must be on sys.path. (Adding an __init__.py there would also
# work but would change test discovery semantics.)
pythonpath = ["tests/unit"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = [
    "E501",   # line length handled by formatter
    "E402",   # module-level code before imports (used in shim patterns)
    "E722",   # bare except tolerated in legacy error handlers
    "E741",   # allow "l" as variable name in data-pipeline loops
    "B006",   # mutable default args in legacy command signatures
    "B007",   # unused loop vars that document intent
    "B904",   # raise-from chaining not uniformly applied
    "F841",   # locally-bound unused vars kept for clarity
    "F821",   # handled via runtime imports in legacy paths
    "W291",   # trailing whitespace (formatter handles)
    "F401",   # unused imports at package boundaries used for re-exports
    "E712",   # explicit `== False` reads OK in some legacy guards
    "B005",   # .strip with multi-char string is intentional in glob-pattern helpers
]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
check_untyped_defs = true
